added icon and windows store pack deploy

+++ Weather App/Models/WindData.cs
@@ -9,6 +9,6 @@ namespace Weather_App.Models
 {
     public class WindData
     {
-        public decimal WindSpeed { get; set; }
+        public float WindSpeed { get; set; }
     }
 }
